substitute utility
A python script that find *sql files within a folder and substitute user
defined patterns within those files. After that creates *SIT.sql files
and save it to _SIT folder.
This utility is useful when promoting a code within environments and you
need to subsitute schemas or databases, ie. QS to SIT...